Move backtest related classes to backtest package
