+++ ta4j-core/src/test/java/org/ta4j/core/BarSeriesManagerTest.java
@@ -36,9 +36,10 @@ import java.util.function.Function;
 import org.junit.Before;
 import org.junit.Test;
 import org.ta4j.core.Trade.TradeType;
-import org.ta4j.core.backtest.trade.TradeExecutionModel;
-import org.ta4j.core.backtest.trade.TradeOnCloseModel;
-import org.ta4j.core.backtest.trade.TradeOnOpenModel;
+import org.ta4j.core.backtest.BarSeriesManager;
+import org.ta4j.core.backtest.TradeExecutionModel;
+import org.ta4j.core.backtest.TradeOnCloseModel;
+import org.ta4j.core.backtest.TradeOnOpenModel;
 import org.ta4j.core.indicators.AbstractIndicatorTest;
 import org.ta4j.core.mocks.MockBarSeries;
 import org.ta4j.core.num.DecimalNum;
